# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CEPS Portal API — referral intake, authorizations, invoices, payments, remittances, vendors, reporting
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /referrals:
    get:
      operationId: listReferrals
      tags: [referrals]
      summary: List referrals (filterable, scoped by role)
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: coordinatorId
          in: query
          schema: { type: string }
        - name: clientId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Referrals
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Referral" }
    post:
      operationId: createReferral
      tags: [referrals]
      summary: Submit referral (coordinator portal form or staff manual entry)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferralInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Referral"
  /referrals/{id}:
    get:
      operationId: getReferral
      tags: [referrals]
      summary: Referral detail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Referral
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Referral"
        "404": { description: Not found }
    patch:
      operationId: updateReferral
      tags: [referrals]
      summary: Update referral (status, notes)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferralUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Referral"
    delete:
      operationId: deleteReferral
      tags: [referrals]
      summary: Delete a referral (staff; blocked if converted to a client)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
        "404": { description: Not found }
        "409": { description: Cannot delete a converted referral }
  /referrals/{id}/send-magic-link:
    post:
      operationId: sendReferralMagicLink
      tags: [referrals]
      summary: (Re)send the parent/guardian e-signature magic link
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MagicLinkResult"
  /signature/{token}:
    get:
      operationId: getSignaturePage
      tags: [signature]
      summary: Load agreement data for a magic-link token (no session required)
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Agreement data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignaturePage"
        "400": { description: Invalid or expired token }
    post:
      operationId: submitSignature
      tags: [signature]
      summary: Submit typed-name e-signature
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignatureInput"
      responses:
        "200":
          description: Signed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
        "400": { description: Invalid or expired token }
...
components:
  schemas:
...
    OkResult:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    MagicLinkResult:
      type: object
      required: [sent]
      properties:
        sent: { type: boolean }
        devLink: { type: ["string", "null"], description: "Development only: the link that would have been emailed" }
...
    ReferralIntakeFields:
      type: object
      description: Structured referral form data (Section 9 field list)
      properties:
        regionalCenterName: { type: string }
        coordinatorName: { type: string }
        coordinatorEmail: { type: string }
        coordinatorPhone: { type: string }
        vendorAcceptsChecks: { type: boolean }
        vendorName: { type: string }
        vendorEmail: { type: string }
        vendorPhone: { type: string }
        vendorContactPerson: { type: string }
        vendorServiceStreet: { type: string }
        vendorServiceCity: { type: string }
        vendorServiceZip: { type: string }
        vendorServiceState: { type: string }
        vendorBillingDifferent: { type: string, enum: [yes, no, unknown] }
        vendorBillingStreet: { type: string }
        vendorBillingCity: { type: string }
        vendorBillingZip: { type: string }
        vendorBillingState: { type: string }
        serviceType: { type: string, enum: [direct_pay_459, reimbursement_024] }
        activityDescription: { type: string }
        serviceStartDate: { type: string }
        serviceEndDate: { type: string }
        posNumber: { type: string }
        posStartDate: { type: string }
        posEndDate: { type: string }
        clientFirstName: { type: string }
        clientLastName: { type: string }
        clientDob: { type: string }
        clientUci: { type: string }
        preferredLanguage: { type: string }
        familyRepName: { type: string }
        clientIsMinor: { type: boolean }
        contactPhone: { type: string }
        contactEmail: { type: string }
        contactStreet: { type: string }
        contactCity: { type: string }
        contactZip: { type: string }
        contactState: { type: string }
    Referral:
      type: object
      required: [id, clientId, referralDate, status]
      properties:
        id: { type: string }
        clientId: { type: string }
        clientName: { type: ["string", "null"] }
        serviceCoordinatorId: { type: ["string", "null"] }
        coordinatorName: { type: ["string", "null"] }
        referralDate: { type: string }
        status:
          type: string
          enum: [intake, pending_signature, pending_auth, pending_w9, pending_invoice, active, closed]
        submittedVia: { type: ["string", "null"], enum: [portal, staff_manual_entry, null] }
        intakeFields: { $ref: "#/components/schemas/ReferralIntakeFields" }
        parentEmail: { type: ["string", "null"] }
        parentSignedAt: { type: ["string", "null"] }
        signedByName: { type: ["string", "null"] }
        altaAuthReceivedAt: { type: ["string", "null"] }
        serviceFrequency: { type: ["string", "null"], enum: [one_time, monthly, null] }
        notes: { type: ["string", "null"] }
        createdAt: { type: ["string", "null"] }
    ReferralInput:
      type: object
      required: [intakeFields]
      properties:
        intakeFields: { $ref: "#/components/schemas/ReferralIntakeFields" }
        submittedVia: { type: string, enum: [portal, staff_manual_entry] }
        parentEmail: { type: string }
        serviceFrequency: { type: string, enum: [one_time, monthly] }
        notes: { type: string }
    ReferralUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [intake, pending_signature, pending_auth, pending_w9, pending_invoice, active, closed]
        serviceCoordinatorId: { type: ["string", "null"] }
        parentEmail: { type: string }
        serviceFrequency: { type: string, enum: [one_time, monthly] }
        notes: { type: string }
        altaAuthReceivedAt: { type: ["string", "null"] }
    SignaturePage:
      type: object
      required: [referralId, clientName, activityDescription, alreadySigned]
      properties:
        referralId: { type: string }
        clientName: { type: string }
        activityDescription: { type: ["string", "null"] }
        vendorName: { type: ["string", "null"] }
        serviceStartDate: { type: ["string", "null"] }
        serviceEndDate: { type: ["string", "null"] }
        serviceType: { type: ["string", "null"] }
        alreadySigned: { type: boolean }
    SignatureInput:
      type: object
      required: [typedName, agreed]
      properties:
        typedName: { type: string, minLength: 1 }
        agreed: { type: boolean }
        createAccount: { type: boolean }
        password: { type: string }
...